Introducing, the MiSAR AIO!
This program automatically handles the process of launching the Parser, Transformation Engine, and Graphical Model Generator. It also handles the job of installing any imports that are not currently present with the system. ( THIS IS A WORK IN PROGRESS SO DON'T EXPECT ALL OF THE BUTTONS TO FUNCTION AS INTENDED )